+++ src/lib/condition.rs
@@ -11,7 +11,9 @@ use crate::environment;
 use crate::profile;
 use crate::scriptengine;
 use crate::types;
-use crate::types::{FlowInfo, RustVersionCondition, ScriptValue, Step, TaskCondition};
+use crate::types::{
+    ConditionScriptValue, FlowInfo, RustVersionCondition, ScriptValue, Step, TaskCondition,
+};
 use crate::version::{is_newer, is_same};
 use envmnt;
 use fsio;
@@ -433,13 +435,24 @@ fn validate_criteria(flow_info: Option<&FlowInfo>, condition: &Option<TaskCondit
     }
 }
 
-fn validate_script(condition_script: &Option<Vec<String>>, script_runner: Option<String>) -> bool {
+pub(crate) fn get_script_text(script: &ConditionScriptValue) -> Vec<String> {
+    match script {
+        ConditionScriptValue::SingleLine(text) => vec![text.clone()],
+        ConditionScriptValue::Text(text) => text.clone(),
+    }
+}
+
+fn validate_script(
+    condition_script: &Option<ConditionScriptValue>,
+    script_runner: Option<String>,
+) -> bool {
     match condition_script {
         Some(ref script) => {
             debug!("Checking task condition script.");
 
+            let script_text = get_script_text(script);
             return scriptengine::invoke_script_pre_flow(
-                &ScriptValue::Text(script.to_vec()),
+                &ScriptValue::Text(script_text),
                 script_runner,
                 None,
                 None,
@@ -458,7 +471,7 @@ pub(crate) fn validate_conditions_without_context(condition: TaskCondition) -> b
 pub(crate) fn validate_conditions(
     flow_info: &FlowInfo,
     condition: &Option<TaskCondition>,
-    condition_script: &Option<Vec<String>>,
+    condition_script: &Option<ConditionScriptValue>,
     script_runner: Option<String>,
 ) -> bool {
     validate_criteria(Some(&flow_info), &condition)