+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Enhancement: Support script runners and shebang in condition scripts #987
 * Enhancement: Support script runner args for rust scripts #1025
+* Enhancement: Support condition_script as string and not just vec (similar to script)
 
 ### v0.37.7 (2024-01-14)